+++ specparam/tests/io/test_utils.py
@@ -1,5 +1,7 @@
 """Tests for specparam.io.utils"""
 
+from pathlib import Path
+
 from specparam.io.utils import *
 
 ###################################################################################################
